added Burger class
The Burger class allows you to create new burgers or create one from those already prepared in the list attached to the overloaded constructor. The next step is to create the Order class, which will allow users to place an order. In the future, all available burgers will probably be accessed through some sort of burger database.