Added "Manger" to each class
Each class now has an additional class whose main goal is to manage the available elements of each class. If a user provides an index that is out of the bounds of the provided list, the element won't be initialized. This feature allows for more efficient element creation and provides cleaner, more readable code. I have fixed a few bugs and already laid my eyes upon a few new ones.